+++ src/Misc/Misc.cs
@@ -233,6 +233,14 @@ namespace mpvnet
 
     public class CommandPaletteItem
     {
+        public CommandPaletteItem() {}
+
+        public CommandPaletteItem(string text, Action action)
+        {
+            Text = text;
+            Action = action;
+        }
+
         public string Text { get; set; } = "";
         public string SecondaryText { get; set; } = "";
         public Action Action { get; set; }